+++ desktop/php/Abeille.php
@@ -74,13 +74,13 @@
     $outils = array(
         'health'    => array( 'bouton'=>'bt_healthAbeille',         'icon'=>'fa-medkit',        'text'=>'{{Santé}}' ),
         'netList'   => array( 'bouton'=>'bt_network',               'icon'=>'fa-sitemap',       'text'=>'{{Réseau}}' ),
-        // 'net'       => array( 'bouton'=>'bt_networkAbeille',        'icon'=>'fa-map',           'text'=>'{{Network Graph}}' ),
+        'net'       => array( 'bouton'=>'bt_networkMap',        'icon'=>'fa-map',           'text'=>'{{Placement réseau (BETA)}}' ),
         'compat'    => array( 'bouton'=>'bt_supportedEqList',       'icon'=>'fa-align-left',    'text'=>'{{Compatibilite}}' ),
         'ota'       => array( 'bouton'=>'bt_Ota',                   'icon'=>'fa-paperclip',     'text'=>'{{Mises-à-jour OTA}}' ),
         'support'   => array( 'bouton'=>'bt_maintenancePage',       'icon'=>'fa-medkit',        'text'=>'{{Maintenance}}' ),
     );
     if (isset($dbgDeveloperMode)) {
-        $outils['net'] = array( 'bouton'=>'bt_networkAbeille', 'icon'=>'fa-map', 'text'=>'{{Network Graph}}' );
+        // $outils['net'] = array( 'bouton'=>'bt_networkMap', 'icon'=>'fa-map', 'text'=>'{{Network Graph}}' );
         $outils['graph'] = array( 'bouton'=>'bt_graph', 'icon'=>'fa-flask', 'text'=>'{{Graph}}' );
     }
 ?>